moved update timestamps back to cli.ts
