+++ cli.ts
@@ -1,6 +1,10 @@
 import { State } from "./cli/state.ts";
-import { Denomander, resolve } from "./deps.ts";
-import { URL_TEMPLATE_BASE, VERSION } from "./consts.ts";
+import { Denomander, format, resolve } from "./deps.ts";
+import {
+  REGEX_MIGRATION_FILE_NAME_LEGACY,
+  URL_TEMPLATE_BASE,
+  VERSION,
+} from "./consts.ts";
 
 /** Initializes Denomander */
 const initDenomander = () => {
@@ -30,6 +34,10 @@ const initDenomander = () => {
     .command(
       "rollback [amount?]",
       "Rolls back migrations. Optional number of rollbacks. If not provided, it will do one.",
+    )
+    .command(
+      "update_timestamps",
+      "Update the timestamp format from milliseconds to timestamp. This command should be run inside of the folder where you store your migrations.",
     );
 
   program.parse(Deno.args);
@@ -52,6 +60,38 @@ const initNessie = async () => {
   await Deno.create(resolve(Deno.cwd(), "db/seeds/.gitkeep"));
 };
 
+const updateTimestamps = () => {
+  const migrationFiles = [...Deno.readDirSync(Deno.cwd())];
+
+  const filteredMigrations = migrationFiles
+    .filter((el) => el.isFile && REGEX_MIGRATION_FILE_NAME_LEGACY.test(el.name))
+    .sort()
+    .map((el) => {
+      const filenameArray = el.name.split("-", 2);
+      const milliseconds = filenameArray[0];
+      const filename = filenameArray[1];
+      const timestamp = new Date(parseInt(milliseconds));
+      const newDateTime = format(timestamp, "yyyyMMddHHmmss");
+
+      return {
+        oldName: el.name,
+        newName: newDateTime + "-" + filename,
+      };
+    });
+
+  filteredMigrations.forEach(({ oldName, newName }) => {
+    Deno.renameSync(oldName, newName);
+  });
+
+  const output = filteredMigrations
+    .map(({ oldName, newName }) => `${oldName} => ${newName}`)
+    .join("\n");
+
+  const encoder = new TextEncoder();
+
+  Deno.stdout.writeSync(encoder.encode(output));
+};
+
 /** Main application */
 const run = async () => {
   try {
@@ -64,6 +104,8 @@ const run = async () => {
 
       if (prog.make) {
         await state.makeMigration(prog.fileName);
+      } else if (prog.update_timestamps) {
+        updateTimestamps();
       } else if (prog["make:seed"]) {
         await state.makeSeed(prog.fileName);
       } else {
